Separated Walk class
Cleaned Walk class up
- put into separate class
